refactor: TIMER_STATUS 상수를 timerConstants.js로 분리
