+++ src/hooks/timer/timerConstants.js
@@ -1,0 +1,7 @@
+export const TIMER_STATUS = {
+  IDLE: "idle",
+  RUNNING: "running",
+  PAUSED: "paused",
+  COMPLETED: "completed",
+  FAILED: "failed",
+};
